Create MarioKartDatabase
As part of my Code First Girls Data and SQL course I completed a project as part of a pair. The task was to create a database with features such as triggers, functions, stored procedures, and events, and query it using joins, subqueries, 'group by' and 'having', and views. We also produced an ER diagram for the database.